Release b.7
This was originally going to be Release b.5, but App Store Connect had some other plans...

now with a reactive VU meter while recording audio!

i am testing for crashes that occur in the Sample Library window.
to reproduce:
1. open sample library window
2. right click on toolbar and select "Icon and Text"
3. click on "hide sidebar" button.

on my machine, this is crashing, and I have no idea why. I may have to disable toolbar customization if this is a reoccurring SwiftUI bug.